django_project/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.1
channels==4.3.1
channels-redis==4.3.0
celery==5.5.3
redis==6.4.0
requests==2.32.5
mysqlclient==2.2.6
django-cors-headers==4.6.0
gunicorn==23.0.0
